Refactor HTML and JS code for readability and maintainability
This commit includes several code changes to improve the readability of the HTML and JS in the index file. It also enhances maintainability by providing clearer component names and restructuring the JavaScript code to more modular functions. Indentation has been added for better visual perception of the nesting levels in the HTML.